@@ -1,18 +1,17 @@
-import { Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { MenuItem } from '../menu-item/menu-item.model';
 import { MEAT_API } from 'app/app.api';
-import { ErrorHandler } from 'app/app.error-handler';
 
 @Injectable()
 export class MenuService {
-  constructor(private http: Http) {
+  constructor(private http: HttpClient) {
   }
 
   menus(id: string): Observable<MenuItem[]>{
-    return this.http.get(`${MEAT_API}/restaurants/${id}/menu`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError);
+    return this
+      .http
+      .get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`);
   }
 }
