@@ -1,13 +1,16 @@
 import { Component, OnInit } from '@angular/core';
 import { Restaurant } from './restaurant/restaurant.model';
 import { RestaurantsService } from './restaurants.service';
-import {
-  trigger,
-  state,
-  style,
-  transition,
-  animate
-} from '@angular/animations';
+import { trigger, state, style, transition, animate } from '@angular/animations';
+import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
+
+import 'rxjs/add/operator/switchMap';
+import 'rxjs/add/operator/do';
+import 'rxjs/add/operator/debounceTime';
+import 'rxjs/add/operator/distinctUntilChanged';
+import 'rxjs/add/operator/catch';
+import 'rxjs/add/observable/from';
+import { Observable } from 'rxjs/Observable';
 
 @Component({
   selector: 'mt-restaurants',
@@ -30,12 +33,31 @@ import {
 export class RestaurantsComponent implements OnInit {
 
   restaurants: Restaurant[];
-
   searchBarState = 'hidden';
 
-  constructor(private restaurantService: RestaurantsService) { }
+  searchForm: FormGroup;
+  searchControl: FormControl;
+
+  constructor(private restaurantService: RestaurantsService,
+    private formBuilder: FormBuilder) { }
 
   ngOnInit() {
+
+    this.searchControl = this.formBuilder.control('');
+
+    this.searchForm = this.formBuilder.group({
+      searchControl: this.searchControl
+    });
+
+    this.searchControl.valueChanges
+      .debounceTime(500)
+      .distinctUntilChanged()
+      .switchMap(searchTerm =>
+        this.restaurantService.restaurants(searchTerm)
+        .catch(() => Observable.from([]))
+    )
+    .subscribe(restaurants => this.restaurants = restaurants);
+
     this.restaurantService.restaurants()
       .subscribe(restaurants => this.restaurants = restaurants);
   }
@@ -45,3 +67,4 @@ export class RestaurantsComponent implements OnInit {
       this.searchBarState === 'hidden' ? 'visible' : 'hidden';
   }
 }
+
